[VendorPatches] Refactor to Composer plugin
The current version of this package is a standalone Symfony Console
application which only works with packages installed in the vendor
directory. This is unfortunate because it prevents usage for packages
installed elsewhere typically managed by the composer/installers
project.

To be able to determine where a packages is installed vendor-patches
needs to retrieve this information from Composer. One way to achieve
this is to implement vendor-patches as a Composer plugin.

This provides access to the installation manager which can provide
the installation directory for a package. This way a lot of the
current logic for determining package directories becomes obsolete.

Also Composer plugins can provide commands and thus vendor-patches no
longer need to work as a standalone Symfony Console application.

Here we update the definition of the package and the central
Generate Command to work as a Composer plugin. Patch detection is now
executed on a per directory level with Composer providing more
detailed package information and more detailed patch information
being returned as well through new value objects.

All obsolete classes and tests are removed and documentation is
updated accordingly.